pyproject.toml
[project]
name = "shiradl"
description = "Download music from YouTube, YouTube Music and Soundcloud, with great metadata and little effort."
requires-python = ">=3.8"
authors =  [{ name = "KraXen72" }, { name = "glomatico" }]
dependencies = [
	"click",
	"yt-dlp",
	"ytmusicapi",
	"mediafile",
	"pillow",
	"requests_cache"
]
readme = "README.md"
dynamic = ["version"]

[project.urls]
repository = "https://github.com/KraXen72/shira"

[build-system]
requires = ["flit_core"]
build-backend = "flit_core.buildapi"

[project.scripts]
shiradl = "shiradl.cli:cli"
